Bump metro version in React Native
Summary: This is the tweak on the release process of metro: after publishing D13075199 and releasing a new version of Metro, this diff upgrades it in RN along with its yarn.lock file. This should allow us to publish new versions of metro without breaking the RN sanity scripts (or disable them).

Reviewed By: hramos

Differential Revision: D13084224

fbshipit-source-id: 19781ae142db11d7dd8b6ca4303366f7ccb32dcc